🚨Changed integers to not break OpenAPI generator
CreateChatCompletionRequest and CreateCompletionRequest had a property called seed with maximums and minimums of 9223372036854776000 and -9223372036854776000, respectively. These were too large of an integer for the OpenAPI generator to handle, so they were changed to the maximums for 32-bit integers in Swift (2^31 - 1 and -2^31).
I forsee no reason this should be a problem, but I suppose it could require switching to a 64-bit integer to handle the large numbers.

Signed-off-by: Evan <evan@brucemcrooster.dev>